firmware: move cpu clock logic to cpu_clock.h|c
