+++ firmware/common/hackrf_core.h
@@ -33,6 +33,7 @@ extern "C" {
 #include "cpld_jtag.h"
 #include "fixed_point.h"
 #include "i2c_bus.h"
+#include "i2c_lpc.h"
 #include "max283x.h"
 #include "max5864.h"
 #include "mixer.h"
@@ -49,6 +50,7 @@ extern "C" {
 #endif
 
 /* TODO: Hide these configurations */
+extern const i2c_lpc_config_t i2c_config_si5351c_fast_clock;
 extern si5351c_driver_t clock_gen;
 extern ssp_config_t ssp_config_w25q80bv;
 
@@ -67,7 +69,6 @@ extern rf_path_t rf_path;
 extern jtag_t jtag_cpld;
 extern i2c_bus_t i2c0;
 
-void cpu_clock_init(void);
 void clock_gen_init(void);
 void clock_gen_shutdown(void);
 void ssp1_set_mode_max283x(void);
